:pencil2: CRLF to LF
Sadly, all files had to be modified. Note that you can tell `git diff`
to ignore this commit (and the LF to CRLF one) with
`--ignore-cr-at-eol`.

Also note that there is an option in git to only commit as LF; it is
prompted on the first installation, and should be found again in `git
config`.

It's not that I don't like CRLF, I'm actually fine with it, but having mixed
CRLF & LF isn't good practice.